pyproject.toml
[project]
name = "shutr7"
version = "0.1.0"
description = "Extract shutter count from Canon R7 camera over USB"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pyusb>=1.2.1",
    "click>=8.1.0",
]

[project.scripts]
shutr7 = "shutr7.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shutr7"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
